Consolidating generators
Prior to this commit, we had two generators touching the same file, with
each generator doing one small thing.

With this commit, I've consolidated some of the logic into a single
generator.

Closes #121
