+++ lib/generators/iiif_print/catalog_controller_generator.rb
@@ -3,11 +3,10 @@ require 'rails/generators'
 
 module IiifPrint
   class CatalogControllerGenerator < Rails::Generators::Base
-    source_root File.expand_path('../templates', __FILE__)
-
     desc "
       This generator makes the following changes to your app:
       1. Adds index fields in CatalogController
+      2. Adjusts Blacklight IIIF Search configuration settings in CatalogController
       "
 
     def add_index_fields_to_catalog_controller
@@ -16,6 +15,18 @@ module IiifPrint
         "\n\n    # IiifPrint index fields\n"\
         "    config.add_index_field 'all_text_tsimv', highlight: true, helper_method: :render_ocr_snippets\n"
       end
+    end
+
+    def adjust_catalog_controller_all_text_config
+      gsub_file('app/controllers/catalog_controller.rb',
+                " full_text_field: 'text',",
+                " full_text_field: 'all_text_tsimv',")
+    end
+
+    def adjust_catalog_controller_is_page_of_config
+      gsub_file('app/controllers/catalog_controller.rb',
+                " object_relation_field: 'is_page_of_s',",
+                " object_relation_field: 'is_page_of_ssim',")
     end
   end
 end
